+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## 0.3.0
 
+### Breaking Changes
+
+- Database schema changes: Add timestamp to indicator results !125
+
+### Other Changes
+
 - Different validation processes of input geometry for entry points (API and CLI) !119
 - Add review process description to contributing guidelines !124
 - Improve report tests by using mocks to avoid querying ohsome API !116